requirements.txt
gitpython
groq
python-dotenv
streamlit
flask